+++ client/src/components/content/programingLanguages.js
@@ -1,0 +1,85 @@
+import Arduino_Svg from "../../images/Programing_Languages/Arduino_Svg.svg";
+import Cplusplus_Svg from "../../images/Programing_Languages/Cplusplus_Svg.svg";
+import Git_Svg from "../../images/Programing_Languages/Git_Svg.svg";
+import Linux_Svg from "../../images/Programing_Languages/Linux_Svg.svg";
+import Css_Svg from "../../images/Programing_Languages/Css_Svg.svg";
+import Html_Svg from "../../images/Programing_Languages/Html_Svg.svg";
+import Js_Svg from "../../images/Programing_Languages/Js_Svg.svg";
+import NodeJs_Svg from "../../images/Programing_Languages/NodeJs_Svg.svg";
+
+import Lader_Svg from "../../images/Programing_Languages/Lader_Svg.svg";
+
+export const programingLanguages = [
+    [
+        {
+            id: 1,
+            title: "Industrial",
+        },
+        {
+            id: 2,
+            title: "",
+
+        },
+        {
+            id: 3,
+            title: "",
+
+        },
+    ],
+    [
+        {
+            title: "Ladder",
+            imgLink: Lader_Svg,
+            website: "https://ladderlogicworld.com/",
+        },
+        /*    {
+               title: "FrequencyInverters",
+               imgLink: FrequencyInverters_Img,
+               website: "",
+           }, */
+        {
+            title: "Arduino",
+            imgLink: Arduino_Svg,
+            website: "https://github.com/rui-oliveira-tech/webSite/",
+        },
+    ],
+    [
+        {
+            title: "C++",
+            imgLink: Cplusplus_Svg,
+            website: "https://www.w3schools.com/cpp/",
+        },
+        {
+            title: "Css",
+            imgLink: Css_Svg,
+            website: "https://www.w3schools.com/css/",
+        },
+        {
+            title: "Html",
+            imgLink: Html_Svg,
+            website: "https://www.w3.org/html/",
+        },
+        {
+            title: "Js",
+            imgLink: Js_Svg,
+            website: "https://developer.mozilla.org/en-US/docs/Web/JavaScript/",
+        },
+        {
+            title: "NodeJs",
+            imgLink: NodeJs_Svg,
+            website: "https://nodejs.org/",
+        },
+    ],
+    [
+        {
+            title: "Git",
+            imgLink: Git_Svg,
+            website: "https://git-scm.com/",
+        },
+        {
+            title: "Linux",
+            imgLink: Linux_Svg,
+            website: "https://www.linux.org/",
+        },
+    ],
+];
